+++ features/scripts/build_android_app.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
 set -o errexit
 
+if [ -z "$FLUTTER_BIN" ]; then
+  FLUTTER_BIN="flutter"
+fi
+
+echo "Flutter Bin: $FLUTTER_BIN"
+
 cd features/fixtures/app
-flutter build apk
+$FLUTTER_BIN build apk