+++ packages/bugsnag_breadcrumbs_dart_io/lib/bugsnag_breadcrumbs_dart_io.dart
@@ -1,0 +1,1 @@
+export 'src/httpclient.dart' show BugsnagHttpClient;